+++ src/app/(main)/like/components/SendLikeSection.tsx
@@ -4,8 +4,8 @@ import NonLikeable from '@public/svg/non-likeable-80.svg';
 
 import { useGetRequestChat } from '@/apis/chat';
 import { useGetSendCrush } from '@/apis/crush';
-import Divider from '@/components/Divider';
-import Spacing from '@/components/Spacing';
+import Divider from '@/components/layout/Divider';
+import Spacing from '@/components/layout/Spacing';
 
 import ChatRequestSection from './ChatRequestSection';
 import CrushPointSection from './CrushPointSection';
