+++ src/app/(sub)/join/funnels/step2/Step2.tsx
@@ -1,4 +1,4 @@
-import Spacing from '@/components/Spacing';
+import Spacing from '@/components/layout/Spacing';
 
 import MoodBoard from './components/MoodBoard';
 import MoodForm from './components/MoodForm';
