+++ src/app/(sub)/join/funnels/step2/components/BirthdaySection.tsx
@@ -3,7 +3,7 @@
 import WaringIcon from '@public/svg/warning-16.svg';
 import { useWatch } from 'react-hook-form';
 
-import Spacing from '@/components/Spacing';
+import Spacing from '@/components/layout/Spacing';
 import { cn } from '@/utils/cn';
 
 import { useJoin1Context } from '../../../components/Join1Context';
