+++ src/app/(sub)/profile/components/KeywordPopup.tsx
@@ -12,7 +12,7 @@ import { Swiper, SwiperSlide } from 'swiper/react';
 import { Button, ButtonWrapper } from '@/components/Button';
 import { Header } from '@/components/Header';
 import { PopupContainer } from '@/components/Overlay';
-import Spacing from '@/components/Spacing';
+import Spacing from '@/components/layout/Spacing';
 import { KEYWORD_LIST } from '@/constants';
 import { useToast } from '@/hooks';
 import { cn } from '@/utils';
