+++ src/components/Popup/Chat/SendChatRequestPopup.tsx
@@ -15,8 +15,8 @@ import { type ChatResponse, usePostChatMessage } from '@/apis/chat';
 import Dday from '@/app/(main)/like/components/Dday';
 import { Header } from '@/components/Header';
 import { PopupContainer } from '@/components/Overlay';
-import Spacing from '@/components/Spacing';
 import TextField from '@/components/TextField';
+import Spacing from '@/components/layout/Spacing';
 import { S3_BASE_URL } from '@/constants';
 import { useGetDistanceFromAddress, useSetCoords, useToast } from '@/hooks';
 import { calculateAge, cn } from '@/utils';
